fixed player refreshing issue
